provider: Add req/resp logging to Orbit
